Rename challenge_creator module to not clash with the app name
